Move SchemaDescription & co.
